Reformat code according to mentor's requirements

+++ app/main.py
@@ -1,3 +1,6 @@
+import os
+
+
 def copy_file(command: str) -> None:
     try:
         command, source_file_name, destination_file_name = command.split()
@@ -5,6 +8,8 @@ def copy_file(command: str) -> None:
         print("Wrong command format")
     else:
         if command == "cp" and source_file_name != destination_file_name:
+            if not os.path.isfile(source_file_name):
+                raise ValueError(f"File '{source_file_name}' doesn't exist")
             with (
                 open(source_file_name) as source_file,
                 open(destination_file_name, "w") as destination_file
